Generating firmwares once again...
I can compile a firmware for the 328. I don't have an Uno
to test, but that might work as well.

Now, I need to get the main tree compiling for AVR again,
so I can recompile the TVM for different processors.

In the meantime, though, I have one or two platforms (possibly)
that I can target with occam programs. That's a win.

The driver could use cleanup, and is fragile, but it's a step.
It is much cleaner than the complex monstrosity that was
plumb1.