Now compiling and uploading.
It turns out bad things happened to FTDI drivers on 10.9. Grr.

This is not yet a cross-platform upload... plumb needs more work.
That, or I need to go back to my IDE, and fix things.

Either way, this will do the command line on Linux and OSX, I think. At least, it works for me on Linux and OSX. And, with two test programs. That's practically bullet-proof testing.

Either way, it's a first step. Now we can compile and upload.