Added in BYOE.
The bring-your-own-editor version "just works." The changes to make the IDE work
were also the majority of the changes necessary to make the BYOE interface work.

Now, I might swing things around so the command-line version relies on the same infrastructure. This would mean that all three versions run from the same codebase... which, would be easier to maintain.