Use local binaries on Mac.
This fixes the IDE so that it has the config to find local configuration when compiled (I think). However, it likely means that I won't be able to run the command-line version correctly... it now uses the local config stuff for avrdude.

That will need to be fixed, perhaps, but hsouldn't be a bit deal.