Was trying to find a bug... ultimately server-side.
From an email I sent to a user, which describes the bug:

Either way, here's your bug:

I decided, in the compile server, to be lazy. Specifically, we merge your compiled bytecode with an ihex of the VM. We used to upload the TVM and your code separately, but for years there was a bug in the 1280/2560 Arduino firmware that broke this behavior (hex loads to the AVR to different locations). So, we started padding out the VM with zeros, and slamming your code into one big file. Thus, every time you hit compile at the moment, you get a VM and your code. Unfortunately, that means you're shipping 20K every time you hit compile, for what is a tiny bytecode compilation otherwise.

The laziness is that I decided to use a pipe to route the merge to a file. Why? I don't know. However:

http://unix.stackexchange.com/questions/11946/how-big-is-the-pipe-buffer

It turns out a pipe has a limit of 2^16 on my Linux VM. And, it just hangs.

I've now simply routed the output of the merge to a file, and read that in for shipping back. Bug fixed.

I had not idea the pipe had a character buffer, or a limit. But, when I saw where it was failing, and the fact that your code pushed the resulting output of raw hex to 65,536 bytes... I became suspicious. Powers of two always make my computer scientist sense tingle...